소영 가계부 - 메인 현월 데이터 없을때 예산&총수입총지출 처리

+++ src/main/java/net/softsociety/secretary/domain/Budget.java
@@ -1,5 +1,7 @@
 package net.softsociety.secretary.domain;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -18,6 +20,6 @@ public class Budget {
 	int familyId;
 	int budgetYear;
 	int budgetMonth;
-	int budgetAmount;
+	BigDecimal budgetAmount;
 	
 }
